St34m Tools/requirements.txt
selenium
subprocess
pillow
signal